[terminal] refactor filesystem as a header
- refactor filesystem as a header
- add terminal.cpp
